code review: move service_windows under subcoomnads/run to unify fx source of truth
