+++ cmd/privateactionrunner/subcommands/run/command.go
@@ -7,6 +7,7 @@
 package run
 
 import (
+	"context"
 	"errors"
 
 	"github.com/DataDog/datadog-agent/comp/core/hostname/remotehostnameimpl"
@@ -36,6 +37,54 @@ type cliParams struct {
 	*command.GlobalParams
 }
 
+// runPrivateActionRunner runs the private action runner with the given configuration and context.
+// This function is shared between the CLI run command and the Windows service.
+func runPrivateActionRunner(ctx context.Context, confPath string, extraConfFiles []string) error {
+	var fxOptions []fx.Option
+
+	// Provide context and setup shutdown listener if running as Windows service
+	if ctx != nil {
+		fxOptions = append(fxOptions,
+			fx.Provide(func() context.Context { return ctx }),
+			fx.Invoke(func(shutdowner fx.Shutdowner) {
+				go func() {
+					<-ctx.Done()
+					_ = shutdowner.Shutdown()
+				}()
+			}),
+		)
+	}
+
+	fxOptions = append(fxOptions,
+		fx.Supply(core.BundleParams{
+			ConfigParams: config.NewAgentParams(confPath, config.WithExtraConfFiles(extraConfFiles)),
+			LogParams:    log.ForDaemon(command.LoggerName, "privateactionrunner.log_file", pkgconfigsetup.DefaultPrivateActionRunnerLogFile)}),
+		core.Bundle(),
+		secretsnoopfx.Module(),
+		fx.Provide(func(c config.Component) settings.Params {
+			return settings.Params{
+				Settings: map[string]settings.RuntimeSetting{
+					"log_level": commonsettings.NewLogLevelRuntimeSetting(),
+				},
+				Config: c,
+			}
+		}),
+		settingsimpl.Module(),
+		remotehostnameimpl.Module(),
+		ipcfx.ModuleReadWrite(),
+		rcserviceimpl.Module(),
+		rcclientimpl.Module(),
+		fx.Supply(rcclient.Params{AgentName: "private-action-runner", AgentVersion: version.AgentVersion}),
+		privateactionrunnerfx.Module(),
+	)
+
+	err := fxutil.Run(fxOptions...)
+	if errors.Is(err, privateactionrunner.ErrNotEnabled) {
+		return nil
+	}
+	return err
+}
+
 // Commands returns a slice of subcommands for the 'private-action-runner' command.
 func Commands(globalParams *command.GlobalParams) []*cobra.Command {
 	cliParams := &cliParams{
@@ -47,32 +96,7 @@ func Commands(globalParams *command.GlobalParams) []*cobra.Command {
 		Short: "Run the Private Action Runner",
 		Long:  `Runs the private-action-runner in the foreground`,
 		RunE: func(_ *cobra.Command, _ []string) error {
-			err := fxutil.Run(
-				fx.Supply(core.BundleParams{
-					ConfigParams: config.NewAgentParams(globalParams.ConfFilePath, config.WithExtraConfFiles(cliParams.ExtraConfFilePath)),
-					LogParams:    log.ForDaemon(command.LoggerName, "privateactionrunner.log_file", pkgconfigsetup.DefaultPrivateActionRunnerLogFile)}),
-				core.Bundle(),
-				secretsnoopfx.Module(),
-				fx.Provide(func(c config.Component) settings.Params {
-					return settings.Params{
-						Settings: map[string]settings.RuntimeSetting{
-							"log_level": commonsettings.NewLogLevelRuntimeSetting(),
-						},
-						Config: c,
-					}
-				}),
-				settingsimpl.Module(),
-				remotehostnameimpl.Module(),
-				ipcfx.ModuleReadWrite(),
-				rcserviceimpl.Module(),
-				rcclientimpl.Module(),
-				fx.Supply(rcclient.Params{AgentName: "private-action-runner", AgentVersion: version.AgentVersion}),
-				privateactionrunnerfx.Module(),
-			)
-			if errors.Is(err, privateactionrunner.ErrNotEnabled) {
-				return nil
-			}
-			return err
+			return runPrivateActionRunner(context.TODO(), globalParams.ConfFilePath, cliParams.ExtraConfFilePath)
 		},
 	}
